fix: adding __bundled endpoint to expose bundled assets for specs making network requests
