+++ packages/server/lib/plugins/preprocessor.coffee
@@ -6,8 +6,7 @@ Promise = require("bluebird")
 appData = require("../util/app_data")
 cwd = require("../cwd")
 plugins = require("../plugins")
-savedState = require("../saved_state")
-resolve = require('./resolve')
+resolve = require("./resolve")
 
 errorMessage = (err = {}) ->
   (err.stack ? err.annotated ? err.message ? err.toString())
@@ -29,9 +28,6 @@ clientSideError = (err) ->
     })
   }())
   """
-
-getOutputPath = (config, filePath) ->
-  appData.projectsPath(savedState.toHashName(config.projectRoot), "bundles", filePath)
 
 baseEmitter = new EE()
 fileObjects = {}
@@ -83,7 +79,7 @@ module.exports = {
       fileObject = fileObjects[filePath] = _.extend(new EE(), {
         filePath,
         shouldWatch,
-        outputPath: getOutputPath(config, baseFilePath)
+        outputPath: appData.getBundledFilePath(config.projectRoot, baseFilePath)
       })
 
       fileObject.on "rerun", ->
