+++ packages/server/lib/routes.js
@@ -3,6 +3,7 @@ const la = require('lazy-ass')
 const check = require('check-more-types')
 const debug = require('debug')('cypress:server:routes')
 
+const AppData = require('./util/app_data')
 const CacheBuster = require('./util/cache_buster')
 const spec = require('./controllers/spec')
 const reporter = require('./controllers/reporter')
@@ -55,6 +56,14 @@ module.exports = ({ app, config, getRemoteState, networkProxy, project, onError 
   // special fallback - serve local files from the project's root folder
   app.get('/__root/*', (req, res) => {
     const file = path.join(config.projectRoot, req.params[0])
+
+    res.sendFile(file, { etag: false })
+  })
+
+  app.get('/__bundled/*', (req, res) => {
+    const file = AppData.getBundledFilePath(config.projectRoot, path.join('src', req.params[0]))
+
+    debug(`Serving dist'd bundle at file path: %o`, { path: file, url: req.url })
 
     res.sendFile(file, { etag: false })
   })
