+++ packages/server/test/unit/saved_state_spec.js
@@ -9,31 +9,6 @@ const appData = require(`${root}lib/util/app_data`)
 const savedState = require(`${root}lib/saved_state`)
 
 describe('lib/saved_state', () => {
-  describe('#toHashName', () => {
-    const projectRoot = '/foo/bar'
-
-    it('starts with folder name', () => {
-      const hash = savedState.toHashName(projectRoot)
-
-      expect(hash).to.match(/^bar-/)
-    })
-
-    it('computed for given path', () => {
-      const hash = savedState.toHashName(projectRoot)
-      const expected = 'bar-1df481b1ec67d4d8bec721f521d4937d'
-
-      expect(hash).to.equal(expected)
-    })
-
-    it('does not handle empty project path', () => {
-      const tryWithoutPath = () => {
-        return savedState.toHashName()
-      }
-
-      expect(tryWithoutPath).to.throw('Missing project path')
-    })
-  })
-
   context('#create', () => {
     beforeEach(() => {
       return savedState.create().then((state) => {
